refactor(EatInOrder): 매장 식사 컨텍스트에서 루트 애그리거트 기준으로 패키징하라
