+++ src/main/java/kitchenpos/eatinorder/order/domain/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorder.domain;
+package kitchenpos.eatinorder.order.domain;
 
 public enum OrderStatus {
     WAITING, ACCEPTED, SERVED, @Deprecated DELIVERING, @Deprecated DELIVERED, COMPLETED
