+++ src/main/java/kitchenpos/eatinorder/order/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorder.domain;
+package kitchenpos.eatinorder.order.domain;
 
 @Deprecated
 public enum OrderType {
