+++ src/main/java/kitchenpos/eatinorder/ordertable/ui/OrderTableRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.eatinorder.ui;
+package kitchenpos.eatinorder.ordertable.ui;
 
-import kitchenpos.eatinorder.application.OrderTableService;
-import kitchenpos.eatinorder.domain.OrderTable;
+import kitchenpos.eatinorder.ordertable.application.OrderTableService;
+import kitchenpos.eatinorder.ordertable.domain.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
