+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,13 +1,13 @@
 package kitchenpos.application;
 
-import kitchenpos.eatinorder.application.EatInOrderService;
-import kitchenpos.eatinorder.domain.EatInOrder;
-import kitchenpos.eatinorder.domain.OrderLineItem;
-import kitchenpos.eatinorder.domain.EatInOrderRepository;
-import kitchenpos.eatinorder.domain.OrderStatus;
-import kitchenpos.eatinorder.domain.OrderTable;
-import kitchenpos.eatinorder.domain.OrderTableRepository;
-import kitchenpos.eatinorder.domain.OrderType;
+import kitchenpos.eatinorder.order.application.EatInOrderService;
+import kitchenpos.eatinorder.order.domain.EatInOrder;
+import kitchenpos.eatinorder.order.domain.OrderLineItem;
+import kitchenpos.eatinorder.order.domain.EatInOrderRepository;
+import kitchenpos.eatinorder.order.domain.OrderStatus;
+import kitchenpos.eatinorder.ordertable.domain.OrderTable;
+import kitchenpos.eatinorder.ordertable.domain.OrderTableRepository;
+import kitchenpos.eatinorder.order.domain.OrderType;
 import kitchenpos.menu.menu.domain.MenuRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
